ENH: Download clang-format 8.0.1 executable
A clang-format executable for all major platforms is downloaded with
development builds with a version compatible with ITK's clang-format
configuration, currently 8.0.1. Git is configured with the executable's
path. The `clangFormat.binary` Git config variable is used to be
compatible with the git-clang-format script provided by LLVM.

In the future, this will help automatically apply ITK's Coding Style
Guidelines in Git hooks and aliases.

For portability, the binaries on Linux are compiled with dockbuild /
centos7-devtoolset4-gcc5, -static-libstdc++. On macOS,
CMAKE_OSX_DEPLOYMENT_TARGET is set to 10.13, the earliest version
support by LLVM 8.0. On Windows, the binary released at llvm.org is
used.