STYLE: Enforce ITK style defined by .clang-format
./Utilities/Maintenance/clang-format.bash \
  --clang-format ~/local/llvm/clang+llvm-8.0.0-x86_64-linux-gnu-ubuntu-18.04/bin/clang-format \
  --tracked

This patch updates the ITK preferred style to one that can be
automatically enforced, integrated with many IDE's, and based
on externally defined documentation and examples.

See: https://discourse.itk.org/t/update-coding-style-for-itk/2055

Most of the changes make inconsistent coding practices consistent
throughout the toolkit.  Community discussions regarding changes
to the current defined coding style are explicitly not included
in this request so that they can be independantly reviewed and
considered.

One large change is changing the {} indentation style from
`Whitesmiths` style https://en.wikipedia.org/wiki/Indentation_style#Whitesmiths_style
to an indentation style consistent with many of the other common
indentation styles.  The primary reason for this change is to
overcome the common deficiency of IDE's and formatters to support
this uncommon style choice.

Many ITK developers provided insightful comments and suggestions that
resulted in this set of style comprimises.

Co-authored-by: Hans Johnson <hans-johnson@uiowa.edu>
Co-authored-by: Dzenan Zukic <dzenan.zukic@kitware.com>
Co-authored-by: Brad King <brad.king@kitware.com>