BUG: operator= defined in [tube|*]SpatialObjectPoint incomplete
The assignment operator for SpatialObjectPoint and its derived classes
that overwrite that member function were not updated to also copy
the SpatialObject pointer maintained within those classes.  That
pointer is critical to computing object to world transforms, and error
messages are given if those transforms are attempted without the
spatial object pointer of the point being set.   This commit adds
the copy of the SpatialObject pointer.