BUG: origin was not accounted for correctly in RayCastInterpolateImageFunction
RayCastInterpolateImageFunction assumes that the bounding box of a
volume is defined by points 0 and itk::Image::Size. Originally, the
image origin was not accounted for but it has been accounted for since
cc94806. However, this patch places the center of the first pixel at
(0,0,0) instead of the corner. Accounting for half a voxel shift fixes
GH-1276. Note, however, that the itk::Image::Direction is still ignored
by this function.