PERF: Use IndexRange in BSplineInterpolationWeightFunction constructor
The constructor of `BSplineInterpolationWeightFunction` originally
created a `tempImage`, just for `ImageRegionConstIteratorWithIndex` to
iterate over an index space of size `m_SupportSize`.

This commit replaces this use of `ImageRegionConstIteratorWithIndex` by
`ZeroBasedIndexRange`, which efficiently iterates over a zero-based
index space, without the need for a `tempImage`.

The #include "itkImageRegionConstIteratorWithIndex.h" is moved from
the HXX file to the test CXX file.