PERF: Use ImageBufferRange in Testing::StretchIntensityImageFilter
Improved the performance of `Testing::StretchIntensityImageFilter`
member function `BeforeThreadedGenerateData()` by replacing its local
`ImageRegionConstIteratorWithIndex` variable by an
`Experimental::ImageBufferRange`.

As a result of this commit, ~40% reduction of the run-time duration of
`filter->Update()` was observed: for an input image of 1024x1024x1024,
the duration dropped from ~2.7 seconds to ~1.6 seconds (using VS2017
Release 64-bit).