ENH: Better reading of Y'CbCR DICOM files
When an Y'CbCR file is encountered by itk::GDCMImageIO, convert to RGB,
and set the pixel type to `RGB` on the output. This default provides the
most commonly desired form of the image. If the native YBR is desired,
gdcmImageIO->SetReadYBRToRGB(false); can be called, and the RGB
conversion will not occur along with a `VECTOR` pixel type provided for
the value read.