ENH: Restored behaviour of depricated Analyze reader w.r.t orientation code
We noticed that after switching from old (ITK3) AnalyzeImageIO reader to
NiftiImageIO reader for Analyze files, the directions have changed
It was caused by the fact that the new NIFTI reader is ignoring
Analyze orientation code, which was breaking our tests.

Code added - more correct interpretation of Analyze header and a Unit Test
that makes sure that orientation is correctly converted to the image directions.