ENH: Separate boundary/non-boundary regions ImageBoundaryFacesCalculator
`NeighborhoodAlgorithm::ImageBoundaryFacesCalculator::operator()` places
the non-boundary region (at the center) in a list together with all the
boundary regions (the faces) that it has found. It appears clearer to
keep the non-boundary region and the boundary regions separate.

This commit adds a `Compute(image, region, radius)` member function
and a result type `Result` to `ImageBoundaryFacesCalculator`, which
keep them separate.

It replaces the old `operator()` call by a call to the new `Compute`, in
`MeanImageFilter`, to shorten the code and improve readability.

`operator()` is now implemented by calling `Execute`. `operator()` still
returns the list that has all regions together, as it did before.