COMP: Fix GTest gcc4.6, Mac10.10 error VNumberOfExpectedPixels deduction
Works around compilation errors (presumably because of compiler bugs) in
itkMeanImageFilterGTest and itkMedianImageFilterGTest, as two compilers
failed to deduce the size of an array that is passed by reference, from
an initializer-list argument.

Ubuntu-gcc4.6-Release:

> note: template argument deduction/substitution failed:
> note: couldn't deduce template parameter 'VNumberOfExpectedPixels'

Mac10.10-AppleClang-dbg-x86_64-static

> note: candidate template ignored: couldn't infer template argument 'VNumberOfExpectedPixels'

The fix consists of using `std::vector` instead. Which has another
advantage (compared to using a reference to a C-style array), that it
can be passed directly to the GoogleTest `EXPECT_EQ` macro.

Reported by Dženan Zukić (@dzenanz)