DOC: More clearly describe foreground/background
The generalalization of "binary" as '0' and 'not zero'
is not used in the "Binary[Dilate|Erode]ImageFilter.

Images with values of 0 or 1 require an explicit setting
of the DilateValue or ErodeValue to 1 in order to
designate 1 as the foreground value to be dilated/eroded.

Expand the documentation to more forcefully convey this
important consideration.