ENH: Avoid alias names for [Fore|Back]groundValue
Avoid using the aliased names for in dilate and erode
filters.  The terms DilateValue and ErodeValue are
can be mis-interpreted as the size of the dilation/erosion
of a binary image, rather than understanding them to
be the single intensity value that defines the
foreground/background objects to be operated on.