STYLE: C++ Rule of Zero for Neighborhood Operators, support nothrow move
Adjusted `NeighborhoodOperator` and eight derived classes according to
the C++ Rule of Zero for special member functions, by removing
used-declared default-constructors, copy-constructors, and assignment
operators.

Implicitly fixed an old bug in the assignment of `LaplacianOperator`, as
it previously did not copy its `m_DerivativeScalings` data member.

This commit effectively makes move-construction and move-assignment of
those classes "noexcept", as is checked in newly added static
assertions, in itkNeighborhoodOperatorTest. It could also improves the
performance of a "move" of such a type of object.