ENH: Wrap GPUAnisotropicSmoothing module
Note: Factory class moved to a separate file due to compile errors
related to unimplemented abstract methods.

Wrap CastImageFilter for GPUImage

  gpu_image = itk.cast_image_filter(cpu_image, ttype=(CPUImageType,
    GPUImageType))

For casting GPUImage to Image and back so all filters do
not need to instantiated with itk::GPUImage.

itk::GPUCastImageFilter brought in from the elastix team,
Denis P. Shamonin and Marius Staring. Division of Image Processing,
Department of Radiology, Leiden, The Netherlands, as contributed in the
Insight Journal article:

  http://hdl.handle.net/10380/3393