BUG: MetaIO format for spatial objects updated to use elementspacing
MetaIO allowed points to be written in scaled space - where that
space was related to physical space via ElementSpacing scale vector.
That pseudo-index space didn't preserve orientation or origin,
so it was of limited use.

With ITKv5, SpatialObjects no longer maintain a notion of index space.
Points exist in "object" space or "World" space.  Points are written
in object space.  So, the flawed/limited notation of ElementSpacing of
MetaIO should be considered on reading, but not used on writing.

This fixes reading (conversion) of MetaIO to spatialobject to preserve
elementSpacing on read and keep it as a unit scalar on write.