BUG: Use Direction When Computing Offsets
Currently, the direction matrix of the input image is ignored
when computing the half-pixel offset in the AddVertex method.
When the direction matrix is identity, the result is correct;
in all other cases, the resulting mesh is shifted relative
to the underlying image data.  This small patch accounts for
the direction matrix in calculating the half-pixel offset, and
additionally adds a simple test for cases where the direction
matrix is non-identity.