ENH: Add VectorImage support to MeanImageFilter
Add support for the runtime length vector images of
VectorImageFilter. The GenerateDataInSubregion method is overloaded to
specially handle VariableLengthVector Images. The specialization
avoids memory allocation multiple times per pixel  from temporary
variables, by pulling variable outside the loop.

When the change of variable scope was applied to scalars a 20%
performance decreases was observed with MSVC, so multiple methods are
used.