BUG: Pass CellData In Triangle Cell Subdivision
Currently, any CellData in the input mesh is not passed to the
subdivided output mesh.  This patch checks if any CellData exists,
and if so, passes that data into the child cells of the subdivided
output mesh.  This patch creates a new method, PassCellData, which
is called immediately following each call to AddFaceTriangle.  Moreover,
the interfaces to SplitTriangleFromOneEdge, SplitTriangleFromTwoEdges,
and SplitTriangleFromThreeEdges have all been extended to take the
cell identifier of the input cell, so that it will be available to pass to
the PassCellData method.  This has also required that the class for
iterative subdivision (IterativeCellSubdivisionQuadEdgeMeshFilter) be
updated.