BUG: Delaunay Cell Data
The DelaunayConformingQuadEdgeMeshFilter currently results in a
decoupling of the cell faces from their corresponding cell data.
This patch attempts to address this by re-assigning cell data
on either side of an edge to the new cell indices following the
edge flip.  In testing, this results in a dramatic improvement
in consistency.  Unfortunately, occasional inconsistencies are
still apparent.