ENH: When using cmake > 3.12 use FindPython3 module
The FindPython3 module is more robust for ensuring that python 3 or greater is found.

The FindPythonInterp and FindPythonLib modules are deprecated
in cmake greater than 3.12.0

The GoogleTest framework uses the deprecated FindPythonInterp, so this
patch uses the FindPython3 module, and then uses that more robust
information to set the FindPythonInterp cmake variables so that
wrapping, google tests, documentation generation, etc.. all use the
same found version of Python 3.