ENH: Refactoring of OrientImageFilter into DICOMOrientImageFilter
This changed the orientation notation to follow that of DICOM. This is
a flip of the directions when compared to OrientImageFilter.

The enums are redefined as new style enum classes in the class scope.

The input transform is always used for the given orientation of the
input image. Only the output/desired orientation can be set.