ENH: more granular progress reporting in a few filters
This commit should serve as an example of how to add granular progress reporting to filters.

The style without ProgressReporter is suitable for cases when there are not too many update events,
e.g. updating is done per scan-line or per object.

The style using ProgressReporter2 is suitable when progress reporting is done per pixel.