PERF: Refactor RelabelComponentImageFilter
Parallelized the computation of the label counting with parallel
reduction of the map. Parallelized applying the output production
step. Modernizes the use of STL data structures and
loops. Stream-lined cases. Save iterators used for searching to
exploit coherence in images.

Updated some type to avoid casts. The filter clearly maps input pixels
to output pixels, so those type are used. And the number of objects
type has been changed to SizeValueType.