ENH: Adding filter and wrapping for casting within SO hierarchy
This new Cast filter allows all of a specific type of children of a
SpatialObjuct to be returned in a list, where the list elements are typed to
specifically match the type requested.    Compare with the existing
GetChildren() function, where the return list's elements are cast to
the top level SpatialObject class.

The modifications to the CMakeLists.txt file in the wrapping directory
allows lists of spatial objects and lists of spatial object points to be
accessed via array indexes.

When wrapped for python, these changes enable the following code:

> castSO = itk.CastSpatialObjectFilter[3].New()
> castSO.SetInput(myGroup)
> tubesList = castSO.GetTubes()
>
> tube = tubesList[0]
>
> print(tube.GetPoints()[0].GetPositionInObjectSpace())