BUG: Prevent Non-Manifold Geometry
This patch addresses a long-standing issue in the Cuberille module
which frequently prevented its use with itk::QuadEdgeMesh, due a
number of cases in which this filter produces non-manifold geometry.
This patch determines, for each 2x2x2 neighborhood surrounding a
vertex, how many face-connected components N are present, and replicates
the relevant vertex N times.  Then, in the vertex map, the bitmask
and the position of the pixel index is used to determine which vertex
to select in the construction of each face.