WIP: Hide GDCM symbols
When compiling ITK with static library and with explicit hidden
visibility:

 -DCMAKE_VISIBILITY_INLINES_HIDDEN:BOOL=ON
 -DCMAKE_C_VISIBILITY_PRESET:STRING=hidden
 -DCMAKE_CXX_VISIBILITY_PRESET:STRING=hidden

numerous GDCM symbols are still export on linux and OSX. GDCM was
always defining it's libraries and several utility libraries as
"default" even when static libraries with explicit hidden visibility
was requested.

This can cause conflicting runtime environment if multiple libraries
or systems are using ITK or GDCM. For example SimpleITK's distribution
compiles with ITK symbols hidden, but GDCM symbols are still in it's
library interface. This cause a conflict when imported into 3D
Slicer.