ENH: Add itk.xarray_from_image and itk.image_from_xarray
These functions convert an itk.Image to and from an xarray.DataArray. The
xarray.DataArray preserves spatial metadata while processing in Python.
xarray enables parallel computing with Dask, machine learning, and
modern seralization with Zarr.

These interfaces are complete but experimental. They are available for
testing but are subject to possible future changes and refinements.