ENH: Require 10 or more commits for .zenodo.json
As discussed in
https://github.com/InsightSoftwareConsortium/ITK/pull/1597,
require 10 or more commits for .zenodo.json addition. This truncated the
long tail, and simplifies the contribution process for new contributors
by avoid the need to add an ORCID iD and affiliation to .zenodo.json. An
"ITK Community Members" creator is added to the end of the .zenodo.json
list to acknowledge the contributors with < 10 commits.