ENH: Prefer to use sform over qform when both are set
When the sform format and qform format are both available,
or if sform representation is a orthonormal matrix and qform
is not set, prefer sform as the representation for the ITK
direction cosine when set as NIFTI_XFORM_SCANNER_ANAT.

This gives a slightly higher precision representation
that is closer to the DICOM representation of the
direction cosines than the qform can represent.  The
more accurate representation is imporant for applications
that must match the original DICOM data physical space
representation.

This also is the approach of other popular tools
for dicom format conversions like the widely used
dcm2niix program.