COMP: Fix extra semi-colon compiler warning
ITK demands that macros end with a semi-colon so when
the macro is expanded to nothing, there is an exra
; that is complained about.

https://stackoverflow.com/questions/28777660/practical-differences-between-do-while-0-and-void0-in-macr/28777865
https://stackoverflow.com/questions/1306611/how-do-i-implement-no-op-macro-or-template-in-c

Define new ITK_STATEMENT_NOOP for consistent re-use in functions.

pragma-push/pop operations can not have ';' at end, so protect with
"// clang-format [off|on]" syntax to ensure consistent formatting
across differnet configuration options.