BUG: Have a single histogram threshold calculator for intermodes threshold
Fix the threshold calculator inheritance issue in
`itk::IntermodesThresholdImageFilter` by having a single histogram
threshold calculator variable.

The class was declaring a histogram threshold calculator under the
`m_IntermodesCalculator` ivar, which was of type
`itk::IntermodesThresholdCalculator` type. At the same time, the class was
inheriting from `itk::HistogramThresholdImageFilter`, which was declaring
its own histogram threshod calculator `m_Calculator`, which was of type
`itk::HistogramThredholdCalculator`.

This was making the `itk::IntermodesThresholdImageFilter` objects to have
two histogram calculator instances, and given that the `GenerateData`
method was implemented in the parent class
(`itk::HistogramThresholdImageFilter`) and making use of the
`m_Calculator` ivar, this was yielding a runtime error.

Fixes the runtime exception error mentioned in [this
comment](https://github.com/InsightSoftwareConsortium/ITK/pull/511#issuecomment-497877457).