BUG: Bump TubeTK to fix use of ArrayFire and installation command
Two major fixes in this bump:
TubeTK was missing specification of ArrayFire paths needed for
outside application to compile if ArrayFire was turned on when
ITK/TubeTK was compiled.

TubeTK attempted to install a file that is no longer included with
TubeTK

TubeTK can now build its applications. Requires building ITK with
TubeTK, building SlicerExecutionModel (which depends on ITK), and
then re-building ITK with it pointing to the build of
SlicerExecutionModel.   Ugly, but it works.