BUG: Cannot apply operator stream::<< to new enum type
On Visual studio compiling ITK with a program that asked
itkRecursiveGaussianImageFilter to print itself would result
in an undefined symbol error during linkage.   That undefined
symbol was for the stream::<< operator applied to GaussianOrderEnum
type.   That usage occured in the PrintSelf function of the
Recursive Guassian class.   Here that operator is replaced by
a switch statement on the enum.