STYLE: Make alias variables to assist with clear documentation
Use alias variable names that more clearly document what the intent
of the workspace ivars that are re-used memory this stage of the processing.
The Per-thread arrays of array accumulators have been reduced into thread 0
at this point, and will be normalized into PDF values by dividing by the total
mass of the joint accumulator PDF.

Also move exception checking for pre-existing states to the top of the function call
( there is no need to do computations before checking for pre-state exception cases).