PERF: Avoid processing entire rows based on mathematical certainties
When either the fixedMarginalPDF or movingMarginalPDF (which are sums of
rows or columns of the 2D joint histogram) then the intersection point
must also be zero.  Both the contribution to the final values of such a point
is no change, so one can avoid much testing and subsequent inner conditional
processing in such cases.