BUG: Fix 'test not run' error thrown when built as ITK remote.
When ITK builds ITKRLEImage as an interal remote module, and
BUILD_SHARED_LIBS = OFF, a 'test not run' error is thrown. Changing the
names of tests and adding them to the TestDriver allows them to run
with BUILD_SHARED_LIB = OFF and not throw any sort of error.

Changing the names of tests is required due to similar tests having the
same definitions in source ITK.