COMP: Pass through CMAKE_SH for Eigen3 external
In CMake/itkExternal_Eigen3.cmake, multiple CMake args are passed, but CMAKE_SH should be as well, if defined. This is necessary when using MinGW because otherwise will get an error depending on your PATH configuration: sh.exe was found in your PATH. Thus, when building ITK, we pass -DCMAKE_SH:bool=OFF, but this then breaks on Eigen3. Related to #66.