BUG: Specify itk package in SWIG Python modules
This consistently sets the __package__ for `itk` package modules and
functions.

This addresses Windows import failures. Swig's c-module import
statements changed with Swig 4.0. For the new syntax, __package__ must
be set and consistent on the imported modules.  To simplify things,
Linux and macOS are also updated to Swig 4.0.
itk.FloatingPointExceptions,
itk.AdaptiveHistogramEqualizationImageFilter is disabled because it on
macOs, Linux with Swig 4.0.1.

The library outputs in the build tree are changed to be consistent with
the package layout.

We use the standard API for ImageDuplicator invocation in the NumPy
bridge.

For the embedded submodules on the C side, we update the module name to include
`itk.` We also use a single PyImport_GetModuleDict statement for
simplicity and performance.

For imports of submodules in Python, we now use the full path. The
"swig" attribute is now a dict used as a namespace instead of a
types.ModuleType.

Overriding the `.New()` is moved from pyBase.i to igenerator.py. It is
only added to classes that define a .New() and have .AddObserver(), i.e.
not abstract base classes. itk.PyCommand is excluded to prevent
.AddObserver recursion that blows the stack.

The lazy module creation uses Python 3.5 tooling to create a module with
__package__ set and load it lazily. i.e.

   importlib.util.spec_from_file_location
   importlib.util.module_from_spec
   module.__loader__.exec_module

Loaded modules now have the package path in their name.