BUG: Major re-write of ImageToArray.py example
After trying the code, it's clear that ImageToArray.py hasn't been
run in a long time.  It imports InsightToolkit and numarray, which
have been obsolete for a long time.  And the ITK-Python API it uses
is obsolete too.

So I rewrote it so that it actually works with the current ITK-Python
API.