BUG: Fix Segfault in Delaunay Filter
In the event that the Delaunay filter attempts to flip a non-internal
edge (such as along the boundary of an open mesh), the EulerOperator
returns a nullptr.  Previously, this filter was not properly checking
whether the returned edge was nullptr when attempting to reassign cell
data, causing a segfault.  Although the test mesh (mushroom.vtk) is
open and contains a boundary edge, these edges were apparently not being
added to the priority queue, presumably due to the regularity of the
cells, and therefore this case was not triggered in the testing suite.

In this patch, the test is modified by adding a small amount of noise
to the mesh vertices, causing boundary edges to be added to the priority
queue and triggering the failure.  Second, this patch only attempts to
reassign cell data if the Euler operator does not return a nullptr,
thus preventing the segfault.