BUG: Incomplete removal of ITK_INSTALL_NO_* variables
Replacing ITK_INSTALL_NO_DEVELOPMENT & ITK_INSTALL_NO_LIBRARIES to
FALSE to make explicit that the development and libraries
can not be skipped for installation.  Use the install COMPONENTS
features of modern cmake to controll what is installed.

Fix the path for nifti header installs to match that of
nifti 3.0.0+ versions.