ENH: Update building internal FFTW version
Recent FFTW builds support cmake building which more
naturually fits with ITK building.

Several variables which were conditional based on
if SYSTEM version or local version were harmonized
so that the behavior of the builds is more consistent
across configuration options.  This also allows
for some simplification of the dependant code.