ENH: Update remote modules
Update remote modules.

Used the `/Utilities/Maintenance/UpdateRemoteModules.sh` script to update
the modules.

Manually remove the double quotes for the commit hash before update for
the the `RTK` and `TubeTK` modules to allow the script to work as
expected.

Take advantage of the commit to do minor manual editions on the `RTK` and
`TubeTK` remote files concerning the style.