BUG: Register Dask serialization functions for NDArrayITKBase
Dask distributed may complain `TypeError: No dispatch for <class 'itkPyBufferPython.NDArrayITKBase'>` due to the missing of a deserialization method for NDArrayITKBase. This commit tries to solve this issue by registering a deserialization method in dask for NDArrayITKBase.

Correspondingly, the test case verifies the serialization and deserialization of a NDArrayITKBase instance.