STYLE: Rename ITK_DISALLOW_COPY_AND_ASSIGN to ITK_DISALLOW_COPY_AND_MOVE
Clarifies that the macro does not just disallow copy and assign, but
also move operations. Note that in this context, the term "move" refers
to both move-construct and move-assign.

With this commit, the old macro name will remain available, as long as
`ITK_FUTURE_LEGACY_REMOVE = OFF` (which is the default).