BUG: Support MRC2014 mode 0 as signed 8-bit integer
Per the MRC2014 format header definition the mode 0 has been clarified
as *signed* 8-bit integer. Previously MRCImageIO assumed mode 0 was
unsigned 8-bit. This looks at (valid) min/max values in the header to
determine if the type must be signed, and then set the types
appropriate to enable reading of modern MRC files.

The writing behavior is not changed.