STYLE: TxtTransformIO Read() no longer copy input file to data string
Before processing the content of its input file,
`TxtTransformIOTemplate::Read()` copied the content to a raw `char`
buffer, which was then inserted into to an output string stream, whose
string was then copied into `data`.

None of this appears necessary, as the content of the file can also be
processed line by line, by a `while (std::getline(in, line))` loop,
according to this commit.