BUG: Fix MatrixOffsetTransformBase SetFixedParameters if too few params
When the argument passed to `MatrixOffsetTransformBase` member function
`SetFixedParameters` would contain less than `NInputDimensions` values,
the member function would have undefined behavior, which would typically
just cause a crash.

This issue affects transforms derived from `MatrixOffsetTransformBase`, including:

    AffineTransform
    AzimuthElevationToCartesianTransform
    CenteredAffineTransform
    Euler2DTransform
    FixedCenterOfRotationAffineTransform
    QuaternionRigidTransform
    Rigid2DTransform
    Rigid3DTransform
    ScalableAffineTransform
    ScaleLogarithmicTransform
    ScaleSkewVersor3DTransform
    ScaleTransform
    ScaleVersor3DTransform
    Similarity2DTransform
    Similarity3DTransform
    VersorRigid3DTransform
    VersorTransform

A crash could occur when reading a transform from a transform file, when
this file has an insufficient number of fixed parameters. It was
encountered when loading a transform within 3DSlicer (version
4.11.20200930).

The fix consists of throwing an exception when the number of fixed
parameters is too small. A unit test is included.