BUG: Avoid itkConfig double import trap
This potential failure mode can be summed up in the following general
guideline: “Never add a package directory, or any directory inside a
package, directly to the Python path”.

The reason that adding nested directories to the Python sys.path is
problematic is that every module in that directory is now potentially
accessible under two different names: as a top level module (since the
directory is on sys.path) and as a submodule of the package (if the
higher level directory containing the package itself is also on
sys.path).

We must import the itkConfig module from python path, not local reference
in order to maintain the desired behavior that the module acts
as a singleton for the module values (i.e. once the
itkConfig module is loaded, we assume it is never loaded again,
thus changed values like itkConfig.LazyLoading persist across
diffent modules).

`import .conf.itkConfig` is a different context than
`import itkConfig`, even if they are the same file. The
LazyLoading and other values may be different in the two contexts.

In order to maintain backward compatibility AND avoid the
double import trap by “Never add a package directory, or any directory
inside a package, directly to the Python path”, the itkConfig.py file
must exist outside of the itk package directory (it *must* be a stand-alone
single file module).

http://python-notes.curiousefficiency.org/en/latest/python_concepts/import_traps.html