STYLE: Minimize cmake configured python files
All of the cmake configured variables should be
contained in the minimal number of files.  Move
cmake configuration options to the itkConfig.py.in
file, and convert itkBuildOptions.py.in to a
static python file.

The itkBuildOptions.py was previously not used
internally, so a simple test has been added to
exercise the codebase in itkBuildOptions.py.