BUG: LazyLoading must be threadsafe
The loading of modules in python *must* occur as a
single atomic transaction in multiprocessing environments (i.e.
the module should only be loaded by one thread).

The LazyLoading of ITK did not treat the loading of
modules as an atomic transaction, and multiple threads
would attempt to load the cascading dependencies out of
order.

The `getattr` override that allows LazyLoading to work
in the case where the module is *not* loaded now blocks
while the first thread completes the delayed module loading.
After the first thread completes module load as an atomic
transaction, the other threads fall through (skip loading)
and return the value requested.

Need to use a recursive lock for thread ownership so that the
first thread can acquire a RLock as often as needed while
recursively processing dependant modules lazy loads.  Other threads need
to wait until this first thread releases the RLock.

WIP: SQME Fixup mutli-threading locks.