BUG: init_docstring with filter instance from the current module
A class may be wrapped with different template parameters in multiple
modules. During module initialization, the *init_docstring() method may
try to reference an instance in another module that has not yet been
instantiated. This is problematic when the currently instantiated module is a
dependency of the to-be-instantiated module.

Addresses the issue:

```
import itk
itk.UnaryFunctorImageFilter
```

When RTK is enabled as discussed in #2064 #2258