ENH: Use extern template and explicit instantiation to speed up tests
Add CMake option `ITK_EXTERN_TEMPLATE_AND_EXPLICIT_INSTANTIATION`
to enable this option.
This is only useful if BUILD_TESTING is ON, and it is ignored in the
install tree.

The idea is to speed up compilation instantiating only once the
templates with the set of types used.

Similar instantiation of selected types are used for python wrappings.
The CMake machinery to generate bindings should be re-used here, but
I have done it manually for the selected classes as a proof-of-work.

Opening this PR to compare test times in CI. Setting the option to ON by
default for this purpose.

This option also enable third-parties using ITK to compile their targets
with: `target_compile_definition(target PUBLIC ITK_EXTERN_TEMPLATE_AND_EXPLICIT_INSTANTIATION)`

And provide their own set of files `itkImageSourceExternTemplate.h`
and the associated `itkImageSourceExternTemplate.cxx` to instantiate the types they use.

CI preliminary results:

```
        Description        |    Windows    |     Linux     |     MacOs     |
============================================================================
Without                    |   1h.59m      |   1h.21m      |    2h.7m      |
----------------------------------------------------------------------------
Image,ISource, ItoIFilter  |   1h.25m      |   1h.15m      |    2h.3m      |
----------------------------------------------------------------------------
+ IFileReader              |   xh.xxm      |   xh.xxm      |    xh.xm      |
----------------------------------------------------------------------------
```

Reading [this SO question](https://stackoverflow.com/questions/61477486/using-extern-template-with-third-party-header-only-library), we might need to wait for C++20 modules for this.

> Unfortunately, there’s no way to avoid these [class] instantiations. An
> explicit instantiation declaration of a class template doesn’t prevent
> (implicit) instantiation of that template; it merely prevents
> instantiating its non-inline, non-template member functions (which is
> often none of them!) because some other translation unit will supply the
> actual function symbols and object code.