ENH: Module configuration exec to import
This is a simplication that continues the integration of WrapITK into
ITK. The use of `exec` is a security faux pas, and the use of importlib
instead helps analysis tools like PyInstaller understand dependencies.
The use of WRAPITK_PYTHON_PATH has not been tested or supported since the
WrapITK integration and is removed. Base on Python test runtimes, this change is also
slightly (1%) faster.